+++ inlang/development-projects/inlang-svelte/src/routes/[lang]/+page.ts
@@ -1,0 +1,3 @@
+export const load = async ({ data }) => {
+	return { ...data, "page.js": 1 }
+}