+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/src/pages/_app.tsx
@@ -1,3 +1,4 @@
+import React from "react"
 import type { AppProps } from "next/app"
 import { ParaglideJS } from "@inlang/paraglide-js-adapter-next/pages"
 import { Header } from "@/lib/Header"
@@ -7,6 +8,7 @@ export default function App({ Component, pageProps }: AppProps) {
 	return (
 		<ParaglideJS>
 			<Header />
+			{/** @ts-ignore */}
 			<Component {...pageProps} />
 		</ParaglideJS>
 	)