+++ inlang/source-code/paraglide/paraglide-js-adapter-vite/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/paraglide-js-adapter-vite
 
+## 1.2.31
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.20
+
 ## 1.2.30
 
 ### Patch Changes